refactor(engine): move LinePrefixer into its own query submodule
The log-prefixing addition pushed query/mod.rs over the per-file line limit.
Extract LinePrefixer (and its tests) into query/log_prefix.rs, bringing
query/mod.rs back under the gate with no behaviour change.
